download chunk and composite actions
attempted to get request, may have to use another library

+++ Core/WindowManager.h
@@ -10,19 +10,19 @@ class cContainer;
 class WindowManager
 {
 public:
-	static void ExecuteActionID(uint16_t actionID);
-	static cContainer* GetWindowByContainerID(uint16_t containerID);
-	static void CreateContainer(uint16_t containerID, cContainer* parent, wxConstraintPosition constraints);
+	static void ExecuteActionID(uint32_t actionID);
+	static cContainer* GetWindowByContainerID(uint32_t containerID);
+	static void CreateContainer(uint32_t containerID, cContainer* parent, wxConstraintPosition constraints);
 	static void DestroyContainer(cContainer* containerWindow);
 	static void SetTopContainer(cContainer* window);
 private:
 	static struct Container* topContainer;
 	static std::map<uint16_t, cContainer*> containerWindows;
 
-	static void LinkContainerIDWithWindow(uint16_t containerID, cContainer* containerWindow);
-	static void UnlinkContainerID(uint16_t containerID);
+	static void LinkContainerIDWithWindow(uint32_t containerID, cContainer* containerWindow);
+	static void UnlinkContainerID(uint32_t containerID);
 	static void executeAction(struct Action* action);
 	static void replaceContainers(cContainer* target, struct Container* replaceWith);
-	static void replaceContainerElementIndexWithContent(uint16_t containerID, uint8_t index, uint16_t contentID);
+	static void replaceContainerElementIndexWithContent(uint32_t containerID, uint8_t index, uint32_t contentID);
 };
